Apply suggestions from code review

+++ components/clp-py-utils/clp_py_utils/s3_utils.py
@@ -314,40 +314,34 @@ def s3_put(s3_config: S3Config, src_file: Path, dest_path: str) -> None:
 
 def s3_delete_objects(s3_config: S3Config, object_keys: Set[str]) -> None:
     """
-    Deletes a set of objects from the S3 bucket using AWS's DeleteObjects operation.
-    The objects to be deleted are identified by keys relative to the `s3_config.key_prefix`.
-    Note: `boto3.delete_objects` is limited to 1000 objects per call. This method divides the input
-    object keys into chunks of 1000 in size and calls `boto3.delete_objects` multiple times until
-    all objects are deleted.
+    Deletes objects from an S3 bucket. The objects are identified by keys relative to
+    `s3_config.key_prefix`.
 
-    :param s3_config: S3 configuration specifying the credentials and the bucket to perform deletion.
+    Note: The AWS S3 `DeleteObjects` API, used by `boto3.client.delete_objects`, supports a maximum
+    of 1,000 objects per request. This method splits the provided keys into batches of up to 1,000
+    and issues multiple delete requests until all objects are removed.
+
+    :param s3_config: The S3 config specifying the credentials and the bucket to perform deletion.
     :param object_keys: The set of object keys to delete, relative to `s3_config.key_prefix`.
     :raises: Propagates `boto3.client`'s exceptions.
     :raises: Propagates `boto3.client.delete_object`'s exceptions.
     """
-
-    def _iterate_by_chunk(key_prefix: str, relative_object_keys: Set[str], chunk_size: int):
-        chunk: List[str] = []
-        for relative_obj_key in relative_object_keys:
-            chunk.append(key_prefix + relative_obj_key)
-            if len(chunk) == chunk_size:
-                yield chunk
-                chunk = []
-
-        if chunk:
-            yield chunk
-
-    def _gen_deletion_config(objects_to_delete: List[str]):
-        return {"Objects": [{"Key": obj_to_delete} for obj_to_delete in objects_to_delete]}
-
     boto3_config = Config(retries=dict(total_max_attempts=3, mode="adaptive"))
     s3_client = _create_s3_client(s3_config, boto3_config)
-
-    MAX_OBJECTS_PER_REQUEST = 1000
-    for chunked_keys in _iterate_by_chunk(
-        s3_config.key_prefix, object_keys, MAX_OBJECTS_PER_REQUEST
-    ):
+    max_num_objects_to_delete_per_batch = 1000
+    objects_to_delete: List[str] = []
+    for relative_obj_key in object_keys:
+        objects_to_delete.append(s3_config.key_prefix + relative_obj_key)
+        if len(objects_to_delete) < max_num_objects_to_delete_per_batch:
+            continue
         s3_client.delete_objects(
             Bucket=s3_config.bucket,
-            Delete=_gen_deletion_config(chunked_keys),
+            Delete=_gen_deletion_config(objects_to_delete),
+        )
+        objects_to_delete = []
+
+    if len(objects_to_delete) != 0:
+        s3_client.delete_objects(
+            Bucket=s3_config.bucket,
+            Delete=_gen_deletion_config(objects_to_delete),
         )
